restructure to reduce import overhead for different roles
