--- llm_defender/__init__.py
@@ -1,49 +1,0 @@
-# Base modules
-from .base.utils import (
-    EngineResponse,
-    validate_numerical_value,
-    timeout_decorator,
-    validate_uid,
-    validate_response_data,
-    validate_signature,
-    sign_data,
-    validate_prompt,
-    validate_validator_api_prompt_output,
-)
-
-from .base.config import ModuleConfig
-
-from .base.protocol import SubnetProtocol
-
-from .base.neuron import BaseNeuron
-
-from .base.engine import BaseEngine
-
-# Core modules
-from .core.validator import SubnetValidator
-
-from .core.miner import SubnetMiner
-
-from .core import SupportedAnalyzers
-
-from .core import (
-    prompt_injection_process,
-    sensitive_information_process,
-    prompt_injection_scoring,
-    prompt_injection_penalty,
-    sensitive_information_scoring,
-    sensitive_information_penalty,
-    TokenClassificationEngine,
-    TextClassificationEngine,
-    PromptInjectionAnalyzer,
-    SensitiveInformationAnalyzer,
-)
-
-# Configuration
-config = ModuleConfig().get_full_config()
-
-# Import wandb handler only if it enabled
-if config["wandb_enabled"] is True:
-    from .base.wandb_handler import WandbHandler
-
-config = ModuleConfig().get_full_config()
