+++ llm_defender/core/miner/sensitive_information/token_classification.py
@@ -1,5 +1,5 @@
 """
-This module implements the base-engine used by the prompt-injection
+This module implements the base-engine used by the sensitive-data
 feature of the llm-defender-subnet.
 """
 from typing import List
@@ -8,29 +8,32 @@ import torch
 from os import path, makedirs
 from transformers import (
     AutoTokenizer,
-    AutoModelForSequenceClassification,
+    AutoModelForTokenClassification,
 )
 from transformers import pipeline
 import bittensor as bt
-import llm_defender as LLMDefender
 
-class TextClassificationEngine(LLMDefender.BaseEngine):
-    """Text classification engine for detecting prompt injection.
+# Import custom modules
+import llm_defender.base as LLMDefenderBase
 
-    This class implements an engine that uses text classification to
-    identity prompt injection attacks. The text classification engine is
+
+class TokenClassificationEngine(LLMDefenderBase.BaseEngine):
+    """Token classification engine for detecting sensitive data exposure.
+
+    This class implements an engine that uses token classification to
+    identity sensitive data exposure. The token classification engine is
     the primary detection method along with the heuristics engine
-    detecting prompt injection attacks.
+    detecting sensitive data exposure.
 
     Whereas the heuristics engine is a collection of specialized
-    sub-engines the text-classification engine focuses on analyzing the
+    sub-engines the token-classification engine focuses on analyzing the
     prompt as a whole and thus has a potential to yield better results
     than the heuristic based approaches.
 
     Attributes:
         prompt:
             A str instance displaying the prompt to be analyzed by the 
-            TextClassificationEngine.
+            TokenClassificationEngine.
         name (from the BaseEngine located at llm_defender/base/engine.py):
             A str instance displaying the name of the engine. 
         cache_dir (from the BaseEngine located at llm_defender/base/engine.py):
@@ -51,37 +54,37 @@ class TextClassificationEngine(LLMDefender.BaseEngine):
 
     Methods:
         __init__():
-            Defines the name and prompt attributes for the TextClassificationEngine 
+            Defines the name and prompt attributes for the TokenClassificationEngine 
             object.
         _calculate_confidence():
             Determines the confidence score for a given prompt being malicious & 
             returns the value which ranges from 0.0 (SAFE) to 1.0 (MALICIOUS).
         _populate_data():
             Returns a dict instance that displays the outputs for the 
-            TextClassificationEngine.
+            TokenClassificationEngine.
         prepare():
             Checks and creates a cache directory if it doesn't exist, then 
             calls initialize() to set up the model and tokenizer.
         initialize():
-            Loads the model and tokenizer used for the TextClassificationEngine.
+            Loads the model and tokenizer used for the TokenClassificationEngine.
         execute():
             This function performs classification of the given prompt to
-            enable it to detect prompt injection. The function returns the
+            enable it to detect sensitive data exposure. The function returns the
             label and score provided by the classifier and defines the class
             attributes based on the outcome of the classifier.
     """
 
-    def __init__(self, prompts: List[str] = None, name: str = "prompt_injection:text_classification"):
+    def __init__(self, prompts: List[str] = None, name: str = "sensitive_info:token_classification"):
         """
-        Initializes the TextClassificationEngine object with the name and prompt attributes.
+        Initializes the TokenClassificationEngine object with the name and prompt attributes.
 
         Arguments:
             prompt:
                 A str instance displaying the prompt to be analyzed by the 
-                TextClassificationEngine.
+                TokenClassificationEngine.
             name:
                 A str instance displaying the name of the engine. Default is
-                'prompt_injection:text_classification'
+                'sensitive_info:token_classification'
 
         Returns:
             None
@@ -90,39 +93,24 @@ class TextClassificationEngine(LLMDefender.BaseEngine):
         self.prompts = prompts
 
     def _calculate_confidence(self):
-        """
-        Determines a confidence value based on the self.output attribute. This
-        value will be 0.0 if the 'outcome' flag in self.output is 'SAFE', 0.5 if 
-        the flag value is 'UNKNOWN', and 1.0 otherwise.
-
-        Arguments:
-            None
-
-        Returns:
-            A float instance representing the confidence score, which is either 
-            0.0, 0.5 or 1.0 depending on the state of the 'outcome' flag in the
-            output attribute.
-        """
         # Determine the confidence based on the score
-        if self.output["outcome"] != "UNKNOWN":
-            if self.output["outcome"] == "SAFE":
-                return 0.0
-            else:
-                return 1.0
-        else:
-            return 0.5
+        if self.output["token_data"]:
+            highest_score_entity = max(self.output["token_data"], key=lambda x: x['score'])
+            return float(highest_score_entity["score"])
+        
+        return 0.0
 
     def _populate_data(self, results):
         """
-        Takes in the results from the text classification and outputs a properly
+        Takes in the results from the token classification and outputs a properly
         formatted dict instance which can later be used to generate a confidence 
         score with the _calculate_confidence() method.
         
         Arguments:
             results:
-                A list instance depicting the results from the text classification 
+                A list instance depicting the results from the token classification 
                 pipeline. The first element in the list (index=0) must be a dict
-                instance contaning the flag 'outcome', and possibly the flag 'score'.
+                instance containing the flag 'outcome', and possibly the flag 'score'.
 
         Returns:
             A dict instance with two flags--the 'outcome' flag is required and will 
@@ -132,8 +120,15 @@ class TextClassificationEngine(LLMDefender.BaseEngine):
             This dict instance is later saved to the output attribute.
         """
         if results:
-            return {"outcome": results[0]["label"], "score": results[0]["score"]}
-        return {"outcome": "UNKNOWN"}
+
+            # Clean extra data
+            for result in results:
+                result.pop("start")
+                result.pop("end")
+                result["score"] = float(result["score"])
+
+            return {"outcome": "ResultsFound", "token_data": results}
+        return {"outcome": "NoResultsFound", "token_data": []}
 
     def prepare(self) -> bool:
         """
@@ -164,7 +159,7 @@ class TextClassificationEngine(LLMDefender.BaseEngine):
 
     def initialize(self):
         """
-        Initializes the model and tokenizer for the TextClassificationEngine.
+        Initializes the model and tokenizer for the TokenClassificationEngine.
 
         Arguments:
             None
@@ -173,9 +168,9 @@ class TextClassificationEngine(LLMDefender.BaseEngine):
             tuple:
                 A tuple instance. The elements of the tuple are, in order:
                     model:
-                        The model for the TextClassificationEngine.
+                        The model for the TokenClassificationEngine.
                     tokenizer:
-                        The tokenizer for the TextClassificationEngine.
+                        The tokenizer for the TokenClassificationEngine.
 
         Raises:
             Exception:
@@ -185,12 +180,12 @@ class TextClassificationEngine(LLMDefender.BaseEngine):
                 The ValueError is raised if the model or tokenizer is empty.
         """
         try:
-            model = AutoModelForSequenceClassification.from_pretrained(
-                "laiyer/deberta-v3-base-prompt-injection", cache_dir=self.cache_dir
+            model = AutoModelForTokenClassification.from_pretrained(
+                "lakshyakh93/deberta_finetuned_pii", cache_dir=self.cache_dir
             )
 
             tokenizer = AutoTokenizer.from_pretrained(
-                "laiyer/deberta-v3-base-prompt-injection", cache_dir=self.cache_dir
+                "lakshyakh93/deberta_finetuned_pii", cache_dir=self.cache_dir
             )
         except Exception as e:
             raise Exception(
@@ -203,10 +198,10 @@ class TextClassificationEngine(LLMDefender.BaseEngine):
         return model, tokenizer
 
     def execute(self, model, tokenizer):
-        """Perform text-classification for the prompt.
+        """Perform token-classification for the prompt.
 
         This function performs classification of the given prompt to
-        enable it to detect prompt injection. The function returns the
+        enable it to detect sensitive data exposure. The function returns the
         label and score provided by the classifier and defines the class
         attributes based on the outcome of the classifier.
 
@@ -222,7 +217,7 @@ class TextClassificationEngine(LLMDefender.BaseEngine):
                 empty when the function is called.
             Exception:
                 The Exception will be raised if a general error occurs during the 
-                execution of the text classification pipeline. This is based on 
+                execution of the token classification pipeline. This is based on 
                 try/except syntax.
         """
 
@@ -230,23 +225,22 @@ class TextClassificationEngine(LLMDefender.BaseEngine):
             raise ValueError("Model or tokenizer is empty")
         try:
             pipe = pipeline(
-                "text-classification",
+                "token-classification",
                 model=model,
                 tokenizer=tokenizer,
-                truncation=True,
-                max_length=512,
                 device=torch.device("cuda" if torch.cuda.is_available() else "cpu"),
             )
-            results = pipe(self.prompts)
+            nested_results = pipe([self.prompts[0], self.prompts[0]], aggregation_strategy="first")
+            results = [item for sublist in nested_results if isinstance(sublist, list) for item in sublist]
         except Exception as e:
             raise Exception(
-                f"Error occurred during text classification pipeline execution: {e}"
+                f"Error occurred during token classification pipeline execution: {e}"
             ) from e
 
         self.output = self._populate_data(results)
         self.confidence = self._calculate_confidence()
 
         bt.logging.debug(
-            f"Text Classification engine executed (Confidence: {self.confidence} - Output: {self.output})"
+            f"Token Classification engine executed (Confidence: {self.confidence} - Output: {self.output})"
         )
         return True
