+++ llm_defender/core/validator/prompt_injection/__init__.py
@@ -1,0 +1,6 @@
+from . import process as prompt_injection_process
+
+from .reward import (
+    scoring as prompt_injection_scoring,
+    penalty as prompt_injection_penalty,
+)
