+++ llm_defender/core/validator/sensitive_information/__init__.py
@@ -1,0 +1,6 @@
+from . import process as sensitive_information_process
+
+from .reward import (
+    scoring as sensitive_information_scoring,
+    penalty as sensitive_information_penalty,
+)
