+++ llm_defender/core/validator/sensitive_information/process.py
@@ -1,5 +1,8 @@
 import bittensor as bt
-import llm_defender as LLMDefender
+
+# Import custom modules
+import llm_defender.base as LLMDefenderBase
+import llm_defender.core.validator as LLMDefenderCore
 
 
 def process_response(
@@ -13,12 +16,12 @@ def process_response(
     responses_invalid_uids,
     responses_valid_uids,
 ):
-    # Get the hotkey and coldkey for the response
+    # Get the hotkey for the response
     hotkey = validator.metagraph.hotkeys[uid]
     coldkey = validator.metagraph.coldkeys[uid]
 
     # Get the default response object
-    response_object = LLMDefender.prompt_injection_scoring.get_response_object(
+    response_object = LLMDefenderCore.sensitive_information_scoring.get_response_object(
         uid,
         hotkey,
         coldkey,
@@ -30,13 +33,13 @@ def process_response(
     )
 
     # Set the score for invalid responses or responses that fail nonce validation to 0.0
-    if not LLMDefender.prompt_injection_scoring.validate_response(
+    if not LLMDefenderCore.sensitive_information_scoring.validate_response(
         hotkey, response.output
     ) or not validator.validate_nonce(response.output["nonce"]):
         bt.logging.debug(f"Empty response or nonce validation failed: {response}")
-        validator.prompt_injection_scores, old_score, unweighted_new_score = (
-            LLMDefender.prompt_injection_scoring.assign_score_for_uid(
-                validator.prompt_injection_scores,
+        validator.scores, old_score, unweighted_new_score = (
+            LLMDefenderCore.sensitive_information_scoring.assign_score_for_uid(
+                validator.scores,
                 uid,
                 validator.neuron_config.alpha,
                 0.0,
@@ -49,16 +52,16 @@ def process_response(
     else:
         response_time = response.dendrite.process_time
 
-        scored_response = calculate_analyzer_score(
+        scored_response = calculate_score(
             prompt, validator, response.output, target, response_time, hotkey
         )
 
-        validator.prompt_injection_scores, old_score, unweighted_new_score = (
-            LLMDefender.prompt_injection_scoring.assign_score_for_uid(
-                validator.prompt_injection_scores,
+        validator.scores, old_score, unweighted_new_score = (
+            LLMDefenderCore.sensitive_information_scoring.assign_score_for_uid(
+                validator.scores,
                 uid,
                 validator.neuron_config.alpha,
-                scored_response["scores"]["binned_analyzer_score"],
+                scored_response["scores"]["total"],
                 query["weight"],
             )
         )
@@ -70,17 +73,17 @@ def process_response(
             "response_time": response_time,
         }
 
-        text_class = [
+        token_class = [
             data
             for data in response.output["engines"]
-            if "text_classification" in data["name"]
+            if "token_classification" in data["name"]
         ]
 
         engine_data = []
 
-        if text_class:
-            if len(text_class) > 0:
-                engine_data.append(text_class[0])
+        if token_class:
+            if len(token_class) > 0:
+                engine_data.append(token_class[0])
 
         responses_valid_uids.append(uid)
 
@@ -93,11 +96,11 @@ def process_response(
         # Populate response data
         response_object["response"] = miner_response
         response_object["engine_data"] = engine_data
-        response_object["analyzer_scored_response"] = scored_response
-        response_object["analyzer_weight_scores"] = {
-            "new": float(validator.prompt_injection_scores[uid]),
+        response_object["scored_response"] = scored_response
+        response_object["weight_scores"] = {
+            "new": float(validator.scores[uid]),
             "old": float(old_score),
-            "change": float(validator.prompt_injection_scores[uid]) - float(old_score),
+            "change": float(validator.scores[uid]) - float(old_score),
             "unweighted": unweighted_new_score,
             "weight": query["weight"],
         }
@@ -113,7 +116,7 @@ def process_response(
                 },
                 {
                     f"{response_object['UID']}:{response_object['hotkey']}_scores_total": response_object[
-                        "analyzer_scored_response"
+                        "scored_response"
                     ][
                         "scores"
                     ][
@@ -122,7 +125,7 @@ def process_response(
                 },
                 {
                     f"{response_object['UID']}:{response_object['hotkey']}_scores_distance": response_object[
-                        "analyzer_scored_response"
+                        "scored_response"
                     ][
                         "scores"
                     ][
@@ -131,7 +134,7 @@ def process_response(
                 },
                 {
                     f"{response_object['UID']}:{response_object['hotkey']}_scores_speed": response_object[
-                        "analyzer_scored_response"
+                        "scored_response"
                     ][
                         "scores"
                     ][
@@ -140,7 +143,7 @@ def process_response(
                 },
                 {
                     f"{response_object['UID']}:{response_object['hotkey']}_raw_scores_distance": response_object[
-                        "analyzer_scored_response"
+                        "scored_response"
                     ][
                         "raw_scores"
                     ][
@@ -149,7 +152,7 @@ def process_response(
                 },
                 {
                     f"{response_object['UID']}:{response_object['hotkey']}_raw_scores_speed": response_object[
-                        "analyzer_scored_response"
+                        "scored_response"
                     ][
                         "raw_scores"
                     ][
@@ -158,21 +161,21 @@ def process_response(
                 },
                 {
                     f"{response_object['UID']}:{response_object['hotkey']}_weight_score_new": response_object[
-                        "analyzer_weight_scores"
+                        "weight_scores"
                     ][
                         "new"
                     ]
                 },
                 {
                     f"{response_object['UID']}:{response_object['hotkey']}_weight_score_old": response_object[
-                        "analyzer_weight_scores"
+                        "weight_scores"
                     ][
                         "old"
                     ]
                 },
                 {
                     f"{response_object['UID']}:{response_object['hotkey']}_weight_score_change": response_object[
-                        "analyzer_weight_scores"
+                        "weight_scores"
                     ][
                         "change"
                     ]
@@ -199,7 +202,7 @@ def process_response(
     return response_object, responses_invalid_uids, responses_valid_uids
 
 
-def calculate_analyzer_score(
+def calculate_score(
     prompt, validator, response, target: float, response_time: float, hotkey: str
 ) -> dict:
     """This function sets the score based on the response.
@@ -210,8 +213,10 @@ def calculate_analyzer_score(
     """
 
     # Calculate distance score
-    distance_score = LLMDefender.prompt_injection_scoring.calculate_subscore_distance(
-        response, target
+    distance_score = (
+        LLMDefenderCore.sensitive_information_scoring.calculate_subscore_distance(
+            response, target
+        )
     )
     if distance_score is None:
         bt.logging.debug(
@@ -220,7 +225,7 @@ def calculate_analyzer_score(
         distance_score = 0.0
 
     # Calculate speed score
-    speed_score = LLMDefender.prompt_injection_scoring.calculate_subscore_speed(
+    speed_score = LLMDefenderCore.sensitive_information_scoring.calculate_subscore_speed(
         validator.timeout, response_time
     )
     if speed_score is None:
@@ -230,25 +235,25 @@ def calculate_analyzer_score(
         speed_score = 0.0
 
     # Validate individual scores
-    if not LLMDefender.validate_numerical_value(
+    if not LLMDefenderBase.validate_numerical_value(
         distance_score, float, 0.0, 1.0
-    ) or not LLMDefender.validate_numerical_value(speed_score, float, 0.0, 1.0):
+    ) or not LLMDefenderBase.validate_numerical_value(speed_score, float, 0.0, 1.0):
         bt.logging.error(
             f"Calculated out-of-bounds individual scores (Distance: {distance_score} - Speed: {speed_score}) for the response: {response} from hotkey: {hotkey}"
         )
-        return LLMDefender.prompt_injection_scoring.get_engine_response_object()
+        return LLMDefenderCore.sensitive_information_scoring.get_engine_response_object()
 
     # Set weights for scores
     score_weights = {"distance": 0.85, "speed": 0.15}
 
     # Get penalty multipliers
     distance_penalty, speed_penalty = get_response_penalties(
-        prompt, validator, response, hotkey
+        validator, response, hotkey
     )
 
     # Apply penalties to scores
     (
-        total_analyzer_raw_score,
+        total_score,
         final_distance_score,
         final_speed_score,
     ) = validator.calculate_penalized_scores(
@@ -257,24 +262,16 @@ def calculate_analyzer_score(
 
     # Validate individual scores
     if (
-        not LLMDefender.validate_numerical_value(
-            total_analyzer_raw_score, float, 0.0, 1.0
-        )
-        or not LLMDefender.validate_numerical_value(
+        not LLMDefenderBase.validate_numerical_value(total_score, float, 0.0, 1.0)
+        or not LLMDefenderBase.validate_numerical_value(
             final_distance_score, float, 0.0, 1.0
         )
-        or not LLMDefender.validate_numerical_value(final_speed_score, float, 0.0, 1.0)
+        or not LLMDefenderBase.validate_numerical_value(final_speed_score, float, 0.0, 1.0)
     ):
         bt.logging.error(
-            f"Calculated out-of-bounds individual scores (Total: {total_analyzer_raw_score} - Distance: {final_distance_score} - Speed: {final_speed_score}) for the response: {response} from hotkey: {hotkey}"
+            f"Calculated out-of-bounds individual scores (Total: {total_score} - Distance: {final_distance_score} - Speed: {final_speed_score}) for the response: {response} from hotkey: {hotkey}"
         )
-        return LLMDefender.prompt_injection_scoring.get_engine_response_object()
-
-    normalized_analyzer_score, binned_analyzer_score = (
-        LLMDefender.prompt_injection_scoring.get_normalized_and_binned_scores(
-            total_analyzer_raw_score
-        )
-    )
+        return LLMDefenderCore.sensitive_information_scoring.get_engine_response_object()
 
     # Log the scoring data
     score_logger = {
@@ -284,29 +281,19 @@ def calculate_analyzer_score(
         "score_weights": score_weights,
         "penalties": {"distance": distance_penalty, "speed": speed_penalty},
         "raw_scores": {"distance": distance_score, "speed": speed_score},
-        "analyzer_scores": {
-            "binned_analyzer_score": binned_analyzer_score,
-            "normalized_analyzer_score": normalized_analyzer_score,
-            "total_analyzer_raw": total_analyzer_raw_score,
+        "final_scores": {
+            "total": total_score,
             "distance": final_distance_score,
             "speed": final_speed_score,
         },
     }
 
-    bt.logging.debug(f"Calculated analyzer score: {score_logger}")
+    bt.logging.debug(f"Calculated score: {score_logger}")
 
-    normalized_analyzer_score, binned_analyzer_score = (
-        LLMDefender.prompt_injection_scoring.get_normalized_and_binned_scores(
-            total_analyzer_raw_score
-        )
-    )
-
-    return LLMDefender.prompt_injection_scoring.get_engine_response_object(
-        normalized_analyzer_score=normalized_analyzer_score,
-        binned_analyzer_score=binned_analyzer_score,
-        total_analyzer_raw_score=total_analyzer_raw_score,
-        final_analyzer_distance_score=final_distance_score,
-        final_analyzer_speed_score=final_speed_score,
+    return LLMDefenderCore.sensitive_information_scoring.get_engine_response_object(
+        total_score=total_score,
+        final_distance_score=final_distance_score,
+        final_speed_score=final_speed_score,
         distance_penalty=distance_penalty,
         speed_penalty=speed_penalty,
         raw_distance_score=distance_score,
@@ -314,7 +301,7 @@ def calculate_analyzer_score(
     )
 
 
-def apply_penalty(prompt, validator, response, hotkey) -> tuple:
+def apply_penalty(validator, response, hotkey) -> tuple:
     """
     Applies a penalty score based on the response and previous
     responses received from the miner.
@@ -332,13 +319,13 @@ def apply_penalty(prompt, validator, response, hotkey) -> tuple:
 
     similarity = base = duplicate = 0.0
     # penalty_score -= confidence.check_penalty(validator.miner_responses["hotkey"], response)
-    similarity += LLMDefender.prompt_injection_penalty.check_similarity_penalty(
+    similarity += LLMDefenderCore.sensitive_information_penalty.check_similarity_penalty(
         uid, validator.miner_responses[hotkey]
     )
-    base += LLMDefender.prompt_injection_penalty.check_base_penalty(
+    base += LLMDefenderCore.sensitive_information_penalty.check_base_penalty(
         uid, validator.miner_responses[hotkey], response
     )
-    duplicate += LLMDefender.prompt_injection_penalty.check_duplicate_penalty(
+    duplicate += LLMDefenderCore.sensitive_information_penalty.check_duplicate_penalty(
         uid, validator.miner_responses[hotkey], response
     )
 
@@ -348,11 +335,11 @@ def apply_penalty(prompt, validator, response, hotkey) -> tuple:
     return similarity, base, duplicate
 
 
-def get_response_penalties(prompt, validator, response, hotkey):
+def get_response_penalties(validator, response, hotkey):
     """This function resolves the penalties for the response"""
 
     similarity_penalty, base_penalty, duplicate_penalty = apply_penalty(
-        prompt, validator, response, hotkey
+        validator, response, hotkey
     )
 
     distance_penalty_multiplier = 1.0
