+++ pyproject.toml
@@ -31,7 +31,6 @@ dependencies = [
     "bittensor==7.0.0",
     "gitpython==3.1.43",
     "python-dotenv==1.0.1",
-    "torch==2.3.0"
 ]
 
 [project.urls]
@@ -39,7 +38,7 @@ homepage = "https://github.com/synapsec-ai/llm-defender-subnet"
 
 [project.optional-dependencies]
 wandb = ["wandb==0.17.0"]
-miner = ["transformers==4.41.2", "datasets==2.19.1"]
+miner = ["transformers==4.41.2", "datasets==2.19.1", "torch==2.3.0"]
 validator = ["scikit-learn==1.5.0", "numpy==1.26.4"]
 
 [tool.setuptools.packages.find]
