+++ scripts/prep.py
@@ -2,16 +2,17 @@
 This script prepares the engines before miner is executed.
 """
 import sys
-import llm_defender as LLMDefender
+
+import llm_defender.core.miner as LLMDefenderCore
 
 def prepare_engines():
     """Prepare the engines"""
     # Prepare text classification engines
-    if not LLMDefender.TextClassificationEngine().prepare():
+    if not LLMDefenderCore.TextClassificationEngine().prepare():
         print("Unable to prepare text classification engine for prompt injection")
         sys.exit(1)
 
-    if not LLMDefender.TokenClassificationEngine().prepare():
+    if not LLMDefenderCore.TokenClassificationEngine().prepare():
         print("Unable to prepare text classification engine for sensitive information")
         sys.exit(1)
     print("Prepared Text Classification engines")
